remove save state code that typecast function pointers to Bit32u for some reason. added code to conditionally call into the dynamic core only IF the dynamic core was compiled in. This fixes several compile issues with x86_64.

+++ src/hardware/ide.cpp
@@ -1135,7 +1135,7 @@ void IDEATAPICDROMDevice::on_mode_select_io_complete() {
 	 * pages */
 
 	scan = sector + 8;
-	fence = sector + MIN(sector_total,AllocationLength);
+	fence = sector + MIN((unsigned int)sector_total,(unsigned int)AllocationLength);
 
 	while ((scan+2) < fence) {
 		unsigned char PAGE = *scan++;
